Store CID ref in ContentAddress instead of StorePath
IPFSRef is used to store a reference to an IPFS store path. This
allows us to use sha256 instead of sha1 for the StorePath.

This can get a little complicated with getting the cid when we just
have a store path. To address this, we have to recursively get compute
the cid of each reference in order to get the content.

Also adds get-cid and ensure-cid, replacing ensure-ca. These are used
to determine the cid you would get from that store path. "full"
content address parsing/rendering is also removed.