Implement (in principle) missing IPFS cases for new ca machinary
All this is untested, but the "UnimplementedErrors" are removes and it
does build at least.

Bummer that we had to do the `queryPathInfo` in content-address.cc that
broke the "content-address.cc doesn't know about the Store class"
abstraction boundary, and also necessitated changing many `const Store
&` to `Store &`. But, perhaps we can later make the call site
responsible by having it instead provide a `PathReferences<IPFSRef>`.