Extend the derived path family of types for dyn deps
We want to be able to express `foo.drv!bar!baz` directly, rather than
merely express it imperatively over time with a second nix invocation
using the results from the first.

The specific syntax is such that is that `foo.drv!bar!{x,y,z}` is
allowed, but `foo.drv!{x,y,z}!bar` is not.

To that end, we create `Single{Derivation,BuiltPath}` types, that are
very similar except instead of having multiple outputs (in a set or
map), they have a single one. This is for everything to the left of the
rightmost `!`.