Adapt schedular to work with computed drvs
To avoid dealing with an optional `drvPath` (because we might not know
it yet) everywhere, make an `OuterDerivationGoal`. This goal just
builds/substitutes the derivation file, and then kicks of a build for
that obtained derivation; in other words it does the chaining of goals
when the drv file is missing (as can already be the case) or computed
(new case).

This also means the `getDerivation` state can be removed from
`DerivationGoal`, which makes the "basic derivation goal" case and drv
file cases closer together.