Fix strings vs bytes, fix non-unicode in `MemorySourceAccessor` JSON
We've been abusing `std::string` for a while now. The need to be correct
about non-unicode in `MemorySourceAccessor` and its JSON format is a
good forcing function.

That is fixed, and in the process a new header with `Bytes` and
`BytesView` (that should feel somewhat Rust-like) is created. Many (but
by no means all!) definitions have also been updated to use strings vs
bytes properly.